refactor(broker): move RoundRobinPartitionDistributor to topology module
