+++ broker/src/main/java/io/camunda/zeebe/broker/partitioning/topology/PartitionDistributionResolver.java
@@ -14,11 +14,11 @@ import io.atomix.primitive.partition.PartitionId;
 import io.camunda.zeebe.broker.partitioning.PartitionManagerImpl;
 import io.camunda.zeebe.broker.partitioning.distribution.FixedPartitionDistributor;
 import io.camunda.zeebe.broker.partitioning.distribution.FixedPartitionDistributorBuilder;
-import io.camunda.zeebe.broker.partitioning.distribution.RoundRobinPartitionDistributor;
 import io.camunda.zeebe.broker.system.configuration.ClusterCfg;
 import io.camunda.zeebe.broker.system.configuration.PartitioningCfg;
 import io.camunda.zeebe.topology.PartitionDistributor;
 import io.camunda.zeebe.topology.StaticConfiguration;
+import io.camunda.zeebe.topology.util.RoundRobinPartitionDistributor;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
